Partial commit of patch #911 the non-intrusive parts.

+++ src/mouse_events.cpp
@@ -744,7 +744,7 @@ void mouse_handler::mouse_motion(int x, int y, const bool browse, bool update)
 	int mx = drag_from_x_; // some default value to prevent unlikely SDL bug
 	int my = drag_from_y_;
 	if (dragging_ && !dragging_started_ && (SDL_GetMouseState(&mx,&my) & SDL_BUTTON_LEFT != 0)) {
-		const double drag_distance = pow(drag_from_x_- mx, 2) + pow(drag_from_y_- my, 2);
+		const double drag_distance = pow((double) (drag_from_x_- mx), 2) + pow((double) (drag_from_y_- my), 2);
 		if (drag_distance > drag_threshold*drag_threshold) {
 			dragging_started_ = true;
 			cursor::set_dragging(true);
